+++ src/main/java/com/erigitic/sql/SQLQuery.java
@@ -44,10 +44,11 @@ public class SQLQuery {
         errorMessage = builder.errorMessage;
         dataSource = builder.dataSource;
 
-        if (builder.update)
+        if (builder.update) {
             executeUpdate();
-        else
+        } else {
             executeQuery();
+        }
     }
 
     public static SQLQuery.Builder builder(DataSource dataSource) {
@@ -60,8 +61,9 @@ public class SQLQuery {
 
             Optional<ResultSet> resultSetOpt = Optional.of(conn.prepareStatement(statement).executeQuery());
 
-            if (resultSetOpt.isPresent())
+            if (resultSetOpt.isPresent()) {
                 resultSet = resultSetOpt.get();
+            }
 
             conn.close();
         } catch (SQLException e) {
@@ -92,8 +94,9 @@ public class SQLQuery {
 
     public boolean recordExists() {
         try {
-            if (resultSet.isBeforeFirst())
+            if (resultSet.isBeforeFirst()) {
                 return true;
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -108,8 +111,9 @@ public class SQLQuery {
      */
     public boolean getBoolean() {
         try {
-            if (resultSet.next())
+            if (resultSet.next()) {
                 return resultSet.getBoolean(1);
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -125,8 +129,9 @@ public class SQLQuery {
      */
     public boolean getBoolean(boolean def) {
         try {
-            if (resultSet.next())
+            if (resultSet.next()) {
                 return resultSet.getBoolean(1);
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -141,8 +146,9 @@ public class SQLQuery {
      */
     public int getInt() {
         try {
-            if (resultSet.next())
+            if (resultSet.next()) {
                 return resultSet.getInt(1);
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -158,8 +164,9 @@ public class SQLQuery {
      */
     public int getInt(int def) {
         try {
-            if (resultSet.next())
+            if (resultSet.next()) {
                 return resultSet.getInt(1);
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -174,8 +181,9 @@ public class SQLQuery {
      */
     public BigDecimal getBigDecimal() {
         try {
-            if (resultSet.next())
-                return resultSet.getBigDecimal(1);
+            if (resultSet.next()) {
+                return resultSet.getBigDecimal(1).max(new BigDecimal(Double.MAX_VALUE));
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -191,8 +199,9 @@ public class SQLQuery {
      */
     public BigDecimal getBigDecimal(BigDecimal def) {
         try {
-            if (resultSet.next())
+            if (resultSet.next()) {
                 return resultSet.getBigDecimal(1);
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -207,8 +216,9 @@ public class SQLQuery {
      */
     public String getString() {
         try {
-            if (resultSet.next())
+            if (resultSet.next()) {
                 return resultSet.getString(1);
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -224,8 +234,9 @@ public class SQLQuery {
      */
     public String getString(String def) {
         try {
-            if (resultSet.next())
+            if (resultSet.next()) {
                 return resultSet.getString(1);
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -266,7 +277,6 @@ public class SQLQuery {
             return this;
         }
 
-        // TODO: Not really needed anymore
         public Builder equals(String val) {
             statement += "='" + val + "'";